Switch to ahash + implement fuse by vtx attrib
Switched hashbrown dependency for a simpler ahash one since std uses
hashbrown anyways, we only really want the fast ahash function.

Also implemented a function to fuse vertices together in a tri- or
quad-mesh based on a given attribute. Vertices for which the attribute
value is equal get fused into one. The positions are combined according
to the given join function and the rest of the attributes take the value
fo the first vertex encountered.