Updated mesh builders + fixed obj loader
Updated the mesh builder module that can now build a dodecahedron and
a subdivided box.

A new MeshBuilder interface is created to make building meshes a bit more
consistent among different builders.

The obj loader is fixed to not outut uvs or normals if not found in the
obj file.